fix: dtype mismatch

+++ src/scope_causal_forcing/pipeline.py
@@ -126,7 +126,7 @@ class CausalForcingPipeline(Pipeline):
             text_encoder_path=text_encoder_path,
             tokenizer_path=tokenizer_path,
         )
-        text_encoder = text_encoder.to(device=device)
+        text_encoder = text_encoder.to(device=device, dtype=torch.bfloat16)
         print(f"Loaded text encoder in {time.time() - start:.3f}s")
 
         # Load VAE
